Bug fixes and some major changes like changing all the floats to double and such

+++ src/main/java/com/nony/servbyte/order/Order.java
@@ -51,7 +51,7 @@ public class Order {
 	private List<OrderItem> items;
 
 	@Column(name = "total_price")
-	private float totalPrice;
+	private double totalPrice;
 
 	@Enumerated(EnumType.STRING)
 	@Column(name = "status")
